fix: set default tags for jobs
It is very common, that we forget to set these tags. As such it is
probably best to just set some default values. If a project wants
different values, they can just override them. That way we shouldn't get
randomly stuck jobs or jobs executing on the wrong runners.

(This also fixes some inconsistency in the rules for the dart jobs and
documents, why we need to copy the rules over in that manner until a
specific gitlab issue is fixed.)