fix: disable QoS
QoS is intended to allow routers to prioritize traffic depending on what
it is used for. This however causes problems with some routers, that
don't always support it correctly. In this case some network stack in
between our runners and Gitlab does not play ball and causes some jobs
to randomly fail. This "fixes" that by disabling QoS in the meantime.